touched up the way the holo and pyhton ask/send the anchors
Also, found that the client Id for making the connection to the MQTT broker was linked to the username. The username we had set up was "test", since if someone else on the test.mosquitto.org broker was using this it would cause errors. So I changed the client id to be something more unique.